+++ CHANGELOG.md
@@ -50,6 +50,10 @@
 
 ### aquatic_http
 
+#### Added
+
+* Support exposing a Prometheus endpoint for metrics
+
 #### Changed
 
 * Don't return any response peers if announce event is stopped
@@ -74,6 +78,7 @@
 #### Added
 
 * Add HTTP health check route when running without TLS
+* Support exposing a Prometheus endpoint for metrics
 
 #### Changed